wire up noop slow rendering detector if FrameMetricsAggregator not available
