+++ splunk-otel-android/src/test/java/com/splunk/rum/ActivityCallbacksTest.java
@@ -47,6 +47,7 @@ public class ActivityCallbacksTest {
     private VisibleScreenTracker visibleScreenTracker;
     private final AppStartupTimer startupTimer = new AppStartupTimer();
     private final AppStateListener appStateListener = mock(AppStateListener.class);
+    private final SlowRenderingDetector slowRenderingDetector = mock(SlowRenderingDetector.class);
 
     @Before
     public void setup() {
@@ -57,7 +58,7 @@ public class ActivityCallbacksTest {
     @Test
     public void appStartup() {
         startupTimer.start(tracer);
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         Activity activity = mock(Activity.class);
@@ -99,7 +100,7 @@ public class ActivityCallbacksTest {
 
     @Test
     public void activityCreation() {
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
         startupAppAndClearSpans(testHarness);
 
@@ -144,7 +145,7 @@ public class ActivityCallbacksTest {
 
     @Test
     public void activityRestart() {
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         startupAppAndClearSpans(testHarness);
@@ -181,7 +182,7 @@ public class ActivityCallbacksTest {
     @Test
     public void activityResumed() {
         when(visibleScreenTracker.getPreviouslyVisibleScreen()).thenReturn("previousScreen");
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         startupAppAndClearSpans(testHarness);
@@ -212,7 +213,7 @@ public class ActivityCallbacksTest {
 
     @Test
     public void activityDestroyedFromStopped() {
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         startupAppAndClearSpans(testHarness);
@@ -243,7 +244,7 @@ public class ActivityCallbacksTest {
 
     @Test
     public void activityDestroyedFromPaused() {
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         startupAppAndClearSpans(testHarness);
@@ -289,7 +290,7 @@ public class ActivityCallbacksTest {
 
     @Test
     public void activityStoppedFromRunning() {
-        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener));
+        ActivityCallbacks activityCallbacks = new ActivityCallbacks(tracer, visibleScreenTracker, startupTimer, singletonList(appStateListener), slowRenderingDetector);
         ActivityCallbackTestHarness testHarness = new ActivityCallbackTestHarness(activityCallbacks);
 
         startupAppAndClearSpans(testHarness);
